fix(backend): optimize concurrency, error logging, and race condition safeguards
- Refactor CaptionHub to use tokio::sync::RwLock with fast-path read locking to eliminate Mutex contention
- Acquire advisory transaction locks during caption chunk insertion to prevent sequence_number ABA race conditions
- Map StorageError::Unconfigured to explicit 400 Bad Request response in upload handlers
- Only ACK Redis stream AI job messages on successful processing
- Add error tracing across resources, questions, participants, and caption handlers
- Fix total_heartbeats type in AttendanceSummary model to u64 to prevent cast wrapping

+++ backend/src/models.rs
@@ -124,8 +124,9 @@ pub struct AttendanceSummary {
     pub participant_count: usize,
     pub verified_count: usize,
     pub provisional_count: usize,
-    pub total_heartbeats: u32,
+    pub total_heartbeats: u64,
 }
+
 
 #[derive(Deserialize)]
 pub struct ClaimGuestParticipationRequest {
